+++ src/main/java/ke/co/suncha/simba/aqua/utils/BillMeta.java
@@ -4,7 +4,7 @@ package ke.co.suncha.simba.aqua.utils;
  * @author Maitha Manyala <maitha.manyala at gmail.com> on 4/24/15.
  */
 public class BillMeta {
-    private Integer units = 0;
+    private Double units = 0.0;
     private Double amount = 0.0;
     private String content = "";
     private Boolean billWaterSale = Boolean.TRUE;
@@ -33,11 +33,11 @@ public class BillMeta {
         this.content = content;
     }
 
-    public Integer getUnits() {
+    public Double getUnits() {
         return units;
     }
 
-    public void setUnits(Integer units) {
+    public void setUnits(Double units) {
         this.units = units;
     }
 }
